requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile --output-file=requirements.txt
#
appdirs==1.4.4
    # via requests-cache
attrs==21.2.0
    # via
    #   cattrs
    #   fiona
    #   requests-cache
basicauth==0.4.1
    # via hdx-python-utilities
beautifulsoup4==4.10.0
    # via hdx-python-utilities
boto3==1.18.62
    # via tabulator
botocore==1.21.62
    # via
    #   boto3
    #   s3transfer
cattrs==1.8.0
    # via requests-cache
cchardet==2.1.7
    # via tabulator
certifi==2021.10.8
    # via
    #   fiona
    #   pyproj
    #   requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via tabulator
charset-normalizer==2.0.7
    # via requests
ckanapi==4.6
    # via hdx-python-api
click==8.0.3
    # via
    #   click-plugins
    #   cligj
    #   fiona
    #   tabulator
click-plugins==1.1.1
    # via fiona
cligj==0.7.2
    # via fiona
colorlog==6.5.0
    # via hdx-python-utilities
cryptography==35.0.0
    # via pyopenssl
decorator==5.1.0
    # via jsonpath-ng
dnspython==2.1.0
    # via email-validator
docopt==0.6.2
    # via
    #   ckanapi
    #   num2words
email-validator==1.1.3
    # via hdx-python-utilities
et-xmlfile==1.1.0
    # via openpyxl
exchangerates==0.3.4
    # via hdx-python-country
fiona==1.8.20
    # via geopandas
geopandas==0.10.1
    # via aa-toolbox (setup.py)
greenlet==1.1.2
    # via sqlalchemy
hdx-python-api==5.3.1
    # via aa-toolbox (setup.py)
hdx-python-country==3.0.2
    # via hdx-python-api
hdx-python-utilities==3.0.2
    # via hdx-python-country
html5lib==1.1
    # via hdx-python-utilities
idna==3.3
    # via
    #   email-validator
    #   requests
ijson==3.1.4
    # via tabulator
inflect==5.3.0
    # via quantulum3
jmespath==0.10.0
    # via
    #   boto3
    #   botocore
jsonlines==2.0.0
    # via tabulator
jsonpath-ng==1.5.3
    # via libhxl
libhxl==4.24.1
    # via hdx-python-country
linear-tsv==1.1.0
    # via tabulator
lxml==4.6.3
    # via exchangerates
munch==2.5.0
    # via fiona
ndg-httpsclient==0.5.1
    # via hdx-python-api
num2words==0.5.10
    # via quantulum3
numpy==1.21.2
    # via
    #   aa-toolbox (setup.py)
    #   pandas
openpyxl==3.0.9
    # via tabulator
pandas==1.3.3
    # via geopandas
ply==3.11
    # via
    #   jsonpath-ng
    #   libhxl
pyasn1==0.4.8
    # via
    #   hdx-python-api
    #   ndg-httpsclient
pycparser==2.20
    # via cffi
pydantic==1.8.2
    # via aa-toolbox (setup.py)
pyopenssl==21.0.0
    # via
    #   hdx-python-api
    #   ndg-httpsclient
pyphonetics==0.5.3
    # via hdx-python-country
pyproj==3.2.1
    # via geopandas
python-dateutil==2.8.2
    # via
    #   botocore
    #   hdx-python-utilities
    #   libhxl
    #   pandas
python-io-wrapper==0.2
    # via libhxl
python-slugify==5.0.2
    # via ckanapi
pytz==2021.3
    # via pandas
quantulum3==0.7.9
    # via hdx-python-api
ratelimit==2.2.1
    # via hdx-python-utilities
requests==2.26.0
    # via
    #   ckanapi
    #   exchangerates
    #   libhxl
    #   requests-cache
    #   requests-file
    #   tabulator
requests-cache==0.8.1
    # via libhxl
requests-file==1.5.1
    # via hdx-python-utilities
ruamel.yaml==0.17.16
    # via hdx-python-utilities
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
s3transfer==0.5.0
    # via boto3
shapely==1.7.1
    # via geopandas
six==1.16.0
    # via
    #   basicauth
    #   ckanapi
    #   exchangerates
    #   fiona
    #   html5lib
    #   jsonpath-ng
    #   linear-tsv
    #   munch
    #   pyopenssl
    #   python-dateutil
    #   requests-file
    #   tabulator
    #   url-normalize
soupsieve==2.2.1
    # via beautifulsoup4
sqlalchemy==1.4.25
    # via tabulator
tabulator[cchardet]==1.53.5
    # via hdx-python-utilities
text-unidecode==1.3
    # via python-slugify
typing-extensions==3.10.0.2
    # via pydantic
unicodecsv==0.14.1
    # via tabulator
unidecode==1.3.2
    # via
    #   libhxl
    #   pyphonetics
url-normalize==1.4.3
    # via requests-cache
urllib3==1.26.7
    # via
    #   botocore
    #   requests
    #   requests-cache
webencodings==0.5.1
    # via html5lib
wheel==0.37.0
    # via libhxl
xlrd==2.0.1
    # via tabulator
xlrd3==1.1.0
    # via libhxl

# The following packages are considered to be unsafe in a requirements file:
# setuptools
